requirements.txt
flask
pyjwt
datetime
uuid
pytest
psycopg2-binary
python-dotenv
